+++ packages/server/src/auth/user.repository.ts
@@ -11,14 +11,7 @@ export class UserRepository extends Repository<User> {
   ) {
     super(User, dataSource.createEntityManager());
   }
-  async createAdminUser() {
-    const adminUser = new User();
-    adminUser.username = 'admin';
-    await this.save(adminUser);
-    console.log('Admin user created successfully.');
 
-    await this.accountRepository.createAccountForAdmin(adminUser);
-  }
   async getUser(userId) {
     return await this.findOne({
       where: { id: userId },
